Removed test relying on x86 in variable
This test could be put back with a fake Device
